+++ src/Utility/BBCode.php
@@ -148,6 +148,17 @@ class BBCode
      */
     public function youtube(string $text): string
     {
-        return preg_replace_callback($this->pattern['youtube'], fn($matches): string => $this->Html->youtube(is_url($matches[1]) ? (Youtube::getId($matches[1]) ?: '') : $matches[1]), $text) ?: '';
+        return preg_replace_callback($this->pattern['youtube'], function (array $matches): string {
+            if (is_url($matches[1])) {
+                $id = Youtube::getId($matches[1]) ?: '';
+                parse_str(parse_url($matches[1], PHP_URL_QUERY) ?: '', $query);
+                if (isset($query['t']) && is_string($query['t'])) {
+                    $id .= '?start=' . $query['t'];
+                }
+            } else {
+                $id = $matches[1];
+            }
+            return $this->Html->youtube($id);
+        }, $text) ?: '';
     }
 }